Change automated build frequency for maint branch
From now on check for changes nightly and rebuild if there was a change.
This is just like the master branch. The weekly restriction was a left over
from the subversion era, where each change in trunk would als trigger a build
in the stable branch. This is no longer the case in git.