Get rid of some of the libfoo.la has moved warnings.
The warning is caused by finding the libtool-lib on a different path than
the one encoded in its libdir statement; it's a purely lexical check,
so if the mingw version, /c/gcdev is given in the -L directive and c:/gcdev
is in the libdir statement, libtool will issue the warning.

The only one remaining is libintl, which is usually found with /c/gcdev/mingw/lib
but whose libdir=/mingw.
Unfortunately just about every other library links to this so it's issued
a lot.