Begin a CMake project to prebuild dependencies.
This first commit builds GLib and Harbuzz with their dependencies.
It will optionally create a pair of tarballs or install the results
into the current ${GLOBAL_DIR}/gnome.

Note that it has modified copies of custom.sh and defaults.sh converted
into cmake as custom.cmake and defaults.cmake. Be sure that those are in
sync with the originals before you use it.