Fix boost detection with autotools
Turns out when BOOST_DIR is set in the environment
the m4 macro looks for boost in BOOST_DIR/staging.
Not what I expected. To have it look in the root
of the installed directory, we need ac_boost_dir. However
our cmake files do look in BOOST_DIR. So I'm simply
setting both here.