Added Croatian-6.0.0.isl
As this is an "unofficial" language, I took the time and translated the newest version. My translation is based upon the current "Croatian-5.3.3.isl" file, with some corrections and additions where needed. I will also send this new file to the person in charge for the "Unofficial translations" at http://www.jrsoftware.org/files/istrans/

Can someone please check if the file's encoding is correct. I work on a Mac-OS, not on Windows. I did convert the file to windows-1250 with a convertor.
If needed I can provide the same file in utf-8.