Adjust modules to allow libchipcard to work with gwenhywfar's pkg-config setup
The problem is pkg-config by default tries to determine a prefix based on
where it finds the .pc file. Unfortunately on Windows that results in a
Windows style path (C:\...). This is relevant here as libchipcard depends
on gwenhywfar paths returned by pkg-config. This is a recent change in
gwenhywfar.

The current libchipcard version depends on this only indirectly as it calls
gwenhwyfar-config, installed by the gwenhywfar package. Future versions of
libchipcard will use pkg-config directly instead.

The fix is to prevent pkg-config from determining the prefix, and just use
the paths as found in the pc file. That can be achieved by setting the
'--dont-define-prefix' pkg-config switch.

As current libchipcard only uses it indirectly via gwenhywfar, this commit
sets it for the gwenhywfar build. However as future versions of libchipcard
will use pkg-config directly the switch is also added for libchipcard
already as well.