inno-setup: remove 'guile.cmd' from installation
The wrapper 'guile.cmd' to be created at install-time effectively sets
the GUILE_LOAD_PATH for guile.  However, the environment for GC at run-time
is implied by the file 'environment' (generated by 'environment.in' at
configure-time, see gnucash repo).  So 'guile.cmd' appears to be obsolete.

Moreover, 'guile.cmd' does not appear in the installation folder, because
the procedure 'MyAfterInstallConfig' is not called at install-time.  Let's
remove the procedure, too.

* inno_setup/gnucash-mingw64.iss:
- remove for 'guile.cmd' creation, run, uninstall commands
- remove procedure 'MyAfterInstallConfig'