Force consistent download filename for signing key
By default wget will append a number to the downloaded file if
a file by that name already exists. This can lead to the wrong
(old) signing key to be passed to pacman if it was updated
upstream and the previous one was still downloaded.
Explicitly setting an output filename will force reuse of that
name. Note this works well in this context only because
we know we're only downloading a single file. It would not
work for a multi-file download.