Extend setup variables to include our version also in AppVersion
Some users pointed out that until now, a version checking tool called
WinGet is not able to find out the installed version of gnucash, as it is
looking for a registry key "DisplayVersion". This might be caused by not
setting the AppVersion value for the inno setup (and there is no other
variable to be found concerning version in inno), which is hence added here.
https://jrsoftware.org/ishelp/index.php?topic=setup_appversion